fix(pwm): cast pwm_timer_index's 0/1 ternary, drop the #109 -Wconversion override
#109 pinned -Wconversion tree-wide but carved out hal/gd32/pwm.c with a
per-source -Wno-conversion override, routing the fix here as a one-line
cast (TODO(#82)). pwm_timer_index()'s `(periph == TIMER0) ? 0u : 1u`
promotes to unsigned int before narrowing into the uint8_t return; both
values fit, but -Wconversion can't prove that from the ternary alone.
Cast explicitly and delete the override now that this source needs it.

+++ hal/gd32/pwm.c
@@ -133,7 +133,13 @@ static uint8_t pwm_align_mode[2];
 /* TIMER base -> pwm_align_mode index. */
 static uint8_t pwm_timer_index(uint32_t periph)
 {
-	return (periph == TIMER0) ? 0u : 1u;
+	/* The ternary's operands are the unsigned-int constants 0u/1u
+	 * (usual arithmetic conversions promote the whole expression to
+	 * unsigned int before the narrowing return), so -Wconversion can't
+	 * see that only 0 or 1 ever reaches this uint8_t return. Both fit;
+	 * the cast is a no-op at runtime, added only to satisfy the warning
+	 * (#109 TODO). */
+	return (uint8_t)((periph == TIMER0) ? 0u : 1u);
 }
 
 int bridge_hw_pwm_set(uint8_t channel, uint32_t period_ns, uint32_t duty_ns)
